requirements-test.txt
-r requirements.txt
django-cms>=3.0
coverage
coveralls
mock>=1.0.1
nose>=1.3.0
django-nose>=1.2
flake8>=2.1.0
tox>=1.7.0

